[Docs] update
Check the documentation and update the localized documents.
Unified the sorting of hotkey functions.
Note: In the current theme, some table formats are set to be centered but actually appear left-aligned. In the new theme that may be used in the future, they will be displayed correctly, but the actual effect is not good. Therefore, those in these tables that do not look good when centered are pre-processed to be left-aligned here to maintain consistency with the original display effect. This will not have any impact on the display of the current document.