Optimize per-frame update logic (#1571)
Splits functionality away from `TechnoClass_AI` /
`TechnoExt::OnEarlyUpdate()` that does not need to be processed for all
types of technos.

- Splits logic concerning `FootClass` only from
`TechnoExt::UpdateTypeData()` to `TechnoExt::UpdateTypeData_Foot()`
called from `FootClass_AI`. While `OpenTopped` is read for all technos,
it does not work correctly for buildings so it was also moved.
- Moves tunnel enter / exit processing and relevant hooks entirely to
appropriate update functions instead of general ones.
- `TechnoExt::DepletedAmmoActions()` is now called from `UnitClass_AI`
(not a new hook - was repurposed/generalized from `KeepTargetOnMove`
feature which was split into its own function too).
- `TechnoExt::UpdateGattlingRateDownReset()` is now called from
`UnitClass_AI` and `BuildingClass_AI`.

Some notes:
- `TechnoExt::EatPassengers()` was kept in `TechnoExt::OnEarlyUpdate()`
because it does work with passenger logic used on buildings such as Bio
Reactor.
- Theoretically under current circumstances the entire
`TechnoExt::UpdateTypeData()` could've been moved to be processed for
`FootClass` only as there are no building type changes in the game, but
I suppose this is the future-proofed approach.